pyproject.toml
[project]
name = "beekeeper"
version = "0.1.0"
description = "Authentication system with OAuth support"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "fastapi>=0.95.2",
    "uvicorn>=0.27.1",
    "sqlalchemy>=2.0.27",
    "pydantic[email]>=1.10.13",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.9",
    "alembic>=1.13.1",
    "python-dotenv>=1.0.1",
    "jinja2>=3.1.3",
    "httpx>=0.26.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.1.2",
    "asyncpg>=0.29.0",
    "email-validator>=2.2.0",
    "beautifulsoup4>=4.13.4",
    "bcrypt>=3.2.2,<4.0.0",
    "requests>=2.32.3",
    "greenlet>=3.2.0",
    "PyJWT>=2.8.0",
    "aiohttp>=3.12.12",
    "google-api-python-client>=2.174.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "resend>=2.10.0",
    "psutil>=7.0.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"] 

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
